Rename bpmn diff view according to schema
