+++ test/e2e/src/pages/bpmnDiffView.ts
